Using EOL extension now, including a .hgeol file to make certain line ending specifications explicit Added missing test case for lua-winreg A number of warnings addressed Adding some helpers to deal with user privileges Some improvements regarding the inclusion of a basic set of Lua scripts in the resource section New version of setvcvars.cmd

+++ 3rdparty/lua/src/modules/lua-winreg/src/winreg.c
@@ -297,10 +297,10 @@ BOOL reg_aux_setvalue(lua_State *L, HKEY hKey, const TCHAR * pszVal, int type, i
 		luaL_Buffer B;
 		luaL_buffinit(L, &B);
 		if(lua_istable(L, i)){
-			int n;
-			int last = lua_objlen(L, i);
+			size_t n;
+			size_t last = lua_objlen(L, i);
 			for (n = 1; n <= last; n++) {
-				lua_rawgeti(L, i, n);
+				lua_rawgeti(L, i, (int)n);
 				luaL_addstring(&B, lua_checkstring(L, -1));
 				luaL_addchar(&B, 0);
 			}
